requirements.txt
playwright-core==1.40.0
boto3==1.34.0